only remove absolute paths if able

+++ makeflow/test/TR_makeflow_linker_abs.sh
@@ -24,7 +24,9 @@ run() {
 
 clean() {
     cd linker
-    rm -rf /tmp/makeflow_test_absolute_path
+    if [ -w /tmp/makeflow_test_absolute_path ]; then
+        rm -rf /tmp/makeflow_test_absolute_path
+    fi
     rm -rf "$out_dir"
     exit 0
 }
